pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dakota-jazz-calendar"
version = "0.1.0"
description = "Scrape Dakota event calendars into MySQL"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "pymysql>=1.1.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
dakota-scraper = "scraper.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
